# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "企業系統整合管理平台",
        "title": "hsmaster SYSTEM API",
        "termsOfService": "https://eirc.app/",
        "contact": {
            "name": "API System Support",
            "url": "https://eirc.app/",
            "email": "mingzong.lyu@gmail.com"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.1"
    },
    "host": "api.eirc.app",
    "basePath": "/",
    "paths": {
...
        "/v1.0/authority/countersign_type": {
            "get": {
                "description": "條件搜尋會簽類型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CountersignType"
                ],
                "summary": "條件搜尋會簽類型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "會簽類型織ID",
                        "name": "CtID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "會簽類型名稱",
                        "name": "CountersignName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "目前頁數,請從1開始帶入",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "一次回傳比數,請從1開始帶入,最高上限20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/countersign_type.List"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "新增會簽類型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CountersignType"
                ],
                "summary": "新增會簽類型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "會簽類型",
                        "name": "*",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/countersign_type.Created"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1.0/authority/countersign_type/{countersign_typeID}": {
            "get": {
                "description": "取得單一會簽類型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CountersignType"
                ],
                "summary": "取得單一會簽類型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "會簽類型ID",
                        "name": "CtID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/countersign_type.Single"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除單一會簽類型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CountersignType"
                ],
                "summary": "刪除單一會簽類型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "會簽類型ID",
                        "name": "CtID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "更新單一會簽類型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CountersignType"
                ],
                "summary": "更新單一會簽類型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "會簽類型ID",
                        "name": "CtID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新會簽類型",
                        "name": "*",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/countersign_type.Updated"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "code.ErrorMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "detailed": {
                    "description": "詳細錯誤內容"
                },
                "message": {
                    "description": "錯誤回傳訊息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
        "code.SuccessfulMessage": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "正確回傳內容"
                },
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
...
        "countersign_type.Created": {
            "type": "object",
            "required": [
                "countersign_name",
                "creater"
            ],
            "properties": {
                "countersign_name": {
                    "description": "專案編號",
                    "type": "string"
                },
                "creater": {
                    "type": "string"
                }
            }
        },
        "countersign_type.List": {
            "type": "object",
            "required": [
                "limit",
                "page"
            ],
            "properties": {
                "countersign_type": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "countersign_name": {
                                "description": "專案編號",
                                "type": "string"
                            },
                            "create_time": {
                                "description": "創建時間",
                                "type": "string"
                            },
                            "creater": {
                                "type": "string"
                            },
                            "ct_id": {
                                "description": "編號",
                                "type": "string"
                            }
                        }
                    }
                },
                "limit": {
                    "description": "筆數(請從1開始帶入,最高上限20)",
                    "type": "integer"
                },
                "page": {
                    "description": "頁數(請從1開始帶入)",
                    "type": "integer"
                },
                "pages": {
                    "description": "總頁數",
                    "type": "integer"
                }
            }
        },
        "countersign_type.Single": {
            "type": "object",
            "properties": {
                "countersign_name": {
                    "description": "專案編號",
                    "type": "string"
                },
                "create_time": {
                    "description": "創建時間",
                    "type": "string"
                },
                "creater": {
                    "type": "string"
                },
                "ct_id": {
                    "description": "編號",
                    "type": "string"
                }
            }
        },
        "countersign_type.Updated": {
            "type": "object",
            "properties": {
                "countersign_name": {
                    "description": "專案編號",
                    "type": "string"
                }
            }
        }
    }
}